Add authentication and link goals with users

+++ backend/controllers/goalController.js
@@ -3,12 +3,14 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require('../models/goalModel')
+const User = require("../models/userModel");
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals =  await Goal.find()
+//   Find only the goals for the logged in user
+  const goals =  await Goal.find( { user: req.user.id })
   if(!goals.length){
     res.status(200).json({ recordCount: goals.length, message: 'No goals added yet.' });
   } else {
@@ -27,7 +29,8 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text, 
+    user: req.user.id
   })
   res.status(200).json({ message: 'Goal added successfully!', goal: goal });
 });
@@ -41,6 +44,22 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Goal not found')
   }
+
+  const user = await User.findById(req.user.id);
+
+ // Check for user
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+//   ensure that logged in user matches goal's user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
   res.status(200).json({ message: `Goal updated successfully!`, updatedGoal: updatedGoal });
 });
@@ -55,6 +74,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Goal not found')
   }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //   ensure that logged in user matches goal's user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const deletedGoal = await Goal.findOneAndDelete({_id: goalId})
  res.status(200).json({ message: `Deletion successful.`, deletedGoal: deletedGoal });
 });
